Update README_en.md
`<a></a>` means that this line still needs to be updated

What do those settings do:
Import Model -> Influence of .sph textures
Import Model -> Influence of .spa textures
Import Motion -> Margin

Also, I didn't update "Other" and everything that comes after it. (Currently, that is just a crappy Google Translate translation)

Seeing as how you apparently added a lot of stuff to the addon, are you going to write a documentation for it? Or, at least add some info the the hover text. (When you hover over a button, you see some text.)